Got Cucumber to work & fixed my Scenario Test
To make it work I had to move the features, step definitions and the test runner as well as add some vscode settings, and actually fix the register endpoint.
I found out we don't need cucumber-spring as implementing that is a hassle - at least 5 new classes for features we are not using.

Also fixed a few things of the others BDD since they had copy pasted from me, but not all. that's out of scope.